[refactor] 장바구니 예외처리를 리펙토링 + 상품 그룹별 클래스를 리펙토링 한다

+++ src/main/java/com/pofol/main/product/cart/CartController.java
@@ -13,6 +13,7 @@ import org.springframework.security.authentication.AnonymousAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
@@ -35,10 +36,10 @@ public class CartController {
 
         try {
             List<CartDto> productCount = cartService.goCartProductCount(cartDtoList);
-            return new ResponseEntity<>(productCount, HttpStatus.OK);
+            return ResponseEntity.ok(productCount);
         } catch (Exception e) {
             e.printStackTrace();
-            return ResponseEntity.badRequest().body(cartDtoList);
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(cartDtoList);
         }
     }
 
@@ -80,12 +81,13 @@ public class CartController {
     // 장바구니에 상품 저장
     @ResponseBody
     @PostMapping("/saveProduct")
+    @Transactional
     public ResponseEntity<String> saveProductCart(@RequestBody List<CartDto> cartDtoList) {
 
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
 
         if(authentication instanceof AnonymousAuthenticationToken){
-            return ResponseEntity.badRequest().body("로그인 이후 이용");
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Not authorized");
         }
 
         for (CartDto cartDto : cartDtoList) {
@@ -94,11 +96,11 @@ public class CartController {
                     cartService.saveCartProduct(cartDto);
                 } catch (Exception e) {
                     e.printStackTrace();
-                    return ResponseEntity.badRequest().body("장바구니 담기 실패");
+                    return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
                 }
             }
         }
-        return ResponseEntity.ok("장바구니 담기 성공");
+        return ResponseEntity.ok().build();
     }
 
     // 장바구니 상품 수량 변경에 따른 가격 변동
@@ -107,10 +109,10 @@ public class CartController {
     public ResponseEntity<CartDto> getCartProductCount(@RequestBody CartDto cartDto) {
         try {
             CartDto cartProductPrice = cartService.getCartProductPrice(cartDto);
-            return new ResponseEntity<>(cartProductPrice, HttpStatus.OK);
+            return ResponseEntity.ok(cartProductPrice);
         } catch (Exception e) {
             e.printStackTrace();
-            return ResponseEntity.badRequest().body(cartDto);
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(cartDto);
         }
     }
 
@@ -118,16 +120,22 @@ public class CartController {
     @ResponseBody
     @PostMapping("/removeProduct")
     public ResponseEntity<String> removeCartProduct(@RequestBody CartDto cartDto) {
-        System.out.println("cartDtoList = " + cartDto);
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        if(authentication instanceof AnonymousAuthenticationToken){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Not authorized");
+        }
+
         try {
 //            for (CartDto cartDto : cartDtoList) {
                 cartService.removeCartProduct(cartDto);
 //            }
         } catch (Exception e) {
             e.printStackTrace();
-            return ResponseEntity.badRequest().body("장바구니 삭제 실패");
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
         }
 
-        return ResponseEntity.ok("장바구니 삭제 성공");
+        return ResponseEntity.ok().build();
     }
 }
